CHORE :: profile componenet 분리 한 번 더 분리
