+++ src/components/Profile/ProfileSection.js
@@ -1,19 +1,14 @@
-import { StyleSheet, Text, TouchableOpacity, View, KeyboardAvoidingView, Platform } from 'react-native';
+import {StyleSheet, Text, TouchableOpacity, View} from 'react-native';
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import axios from "axios";
-import { baseUrl } from "../../utils/api/urls";
-import { useState } from 'react';
+import {baseUrl} from "../../utils/api/urls";
+import {useState} from 'react';
 import Toggle from "../common/input/Toggle/Toggle";
-import { deleteAccessToken, deleteRefreshToken } from "../../utils/storage/token";
-import InputFrame from '../common/frame/InputFrame/InputFrame';
-import BasicInput from '../common/input/BasicInput/BasicInput';
+import {deleteAccessToken, deleteRefreshToken} from "../../utils/storage/token";
 
-export default function ProfileSection({ setLogout, isEditing }) {
+export default function ProfileSection({setLogout}) {
     const [isNoticeAllow, setIsNoticeAllow] = useState(false);
-    const [nickname, setNickname] = useState("");
-    const [password, setPassword] = useState("");
-    const [newPassword, setNewPassword] = useState("");
-    const [reNewPassword, setReNewPassword] = useState("");
+
     const logoutUser = async () => {
         try {
             const accessToken = await AsyncStorage.getItem("access-token");
@@ -31,81 +26,21 @@ export default function ProfileSection({ setLogout, isEditing }) {
     }
     return (
         <>
-            {isEditing
-                ? <KeyboardAvoidingView
-                    style={editStyles.Edit}
-                    behavior={Platform.OS === 'ios' ? 'padding' : 'height'}
-                >
-                    <InputFrame
-                        style={editStyles.Caption}
-                        title="닉네임"
-                        child={
-                            <BasicInput
-                                style={editStyles.Input}
-                                value={nickname}
-                                setValue={setNickname}
-                                placeholder="nickname123"
-                            />
-                        }
+            <View style={styles.Settings}>
+                <View style={styles.NoticeAllow}>
+                    <Text style={styles.NoticeAllowText}>앱 내 알림 허용</Text>
+                    <Toggle
+                        value={isNoticeAllow}
+                        setValue={setIsNoticeAllow}
                     />
-                    <View style={editStyles.passwordInputs}>
-                        <InputFrame
-                            style={editStyles.Caption}
-                            title="현재 비밀번호"
-                            child={
-                                <BasicInput
-                                    style={editStyles.Input}
-                                    value={password}
-                                    setValue={setPassword}
-                                    placeholder="••••••••"
-                                    secureTextEntry
-                                />
-                            }
-                        />
-                        <InputFrame
-                            style={editStyles.Caption}
-                            title="비밀번호 수정"
-                            child={
-                                <BasicInput
-                                    style={editStyles.Input}
-                                    value={newPassword}
-                                    setValue={setNewPassword}
-                                    placeholder="••••••••"
-                                    secureTextEntry
-                                />
-                            }
-                        />
-                        <InputFrame
-                            style={editStyles.Caption}
-                            title="비밀번호 재입력"
-                            child={
-                                <BasicInput
-                                    style={editStyles.Input}
-                                    value={reNewPassword}
-                                    setValue={setReNewPassword}
-                                    placeholder="••••••••"
-                                    secureTextEntry
-                                />
-                            }
-                        />
-                    </View>
-                </KeyboardAvoidingView>
-                : <View style={styles.Settings}>
-                    <View style={styles.NoticeAllow}>
-                        <Text style={styles.NoticeAllowText}>앱 내 알림 허용</Text>
-                        <Toggle
-                            value={isNoticeAllow}
-                            setValue={setIsNoticeAllow}
-                        />
-                    </View>
-                    <TouchableOpacity style={styles.Logout} onPress={logoutUser}>
-                        <Text>로그아웃</Text>
-                    </TouchableOpacity>
-                    <TouchableOpacity onPress={() => console.log('계정 탈퇴')}>
-                        <Text>계정 탈퇴</Text>
-                    </TouchableOpacity>
                 </View>
-            }
+                <TouchableOpacity style={styles.Logout} onPress={logoutUser}>
+                    <Text>로그아웃</Text>
+                </TouchableOpacity>
+                <TouchableOpacity onPress={() => console.log('계정 탈퇴')}>
+                    <Text>계정 탈퇴</Text>
+                </TouchableOpacity>
+            </View>
         </>
     );
 }
@@ -130,18 +65,3 @@ const styles = StyleSheet.create({
     }
 })
 
-const editStyles = StyleSheet.create({
-    Edit: {
-        paddingHorizontal: 45,
-        flex: 1
-    },
-    Caption: {
-        fontSize: 12
-    },
-    passwordInputs: {
-        marginTop: 46
-    },
-    Input: {
-        fontSize: 12
-    }
-})
